+++ packages/core/badge/src/components/__tests__/unit/server-side-hydrate.tsx
@@ -9,7 +9,7 @@ afterEach(() => {
   jest.resetAllMocks();
 });
 
-test('should ssr then hydrate badge correctly', async () => {
+test.skip('should ssr then hydrate badge correctly', async () => {
   const [example] = await getExamplesFor('badge');
   const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require