+++ packages/core/drawer/src/index.js
@@ -6,6 +6,7 @@ export {
   DrawerSkeletonHeader,
   DrawerSkeletonItem,
   DrawerItemGroup,
+  DrawerItem,
 } from './components';
 
 export * from './components/types';