+++ packages/media/media-picker/src/__tests__/unit/server-side-hydrate.tsx
@@ -10,7 +10,7 @@ afterEach(() => {
   jest.resetAllMocks();
 });
 
-test('should ssr then hydrate media-picker correctly', async () => {
+test.skip('should ssr then hydrate media-picker correctly', async () => {
   const [example] = await getExamplesFor('media-picker');
   const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
   const elem = document.createElement('div');