+++ packages/search/global-search/src/components/jira/JiraSearchResultsMapper.ts
@@ -1,7 +1,6 @@
 import {
   ResultsGroup,
   JiraResultsMap,
-  GenericResultMap,
   Result,
   ResultType,
   AnalyticsType,
@@ -23,9 +22,10 @@ const DEFAULT_MAX_OBJECTS = 8;
 const MAX_CONTAINERS = 6;
 const MAX_PEOPLE = 3;
 
-const DEFAULT_JIRA_RESULTS_MAP: GenericResultMap = {
-  objects: [] as Result[],
+const DEFAULT_JIRA_RESULTS_MAP: JiraResultsMap = {
+  objects: [],
   containers: [],
+  people: [],
 };
 
 const isEmpty = (arr: Array<any> = []) => !arr.length;
@@ -36,7 +36,7 @@ const hasNoResults = (
   containers: Array<Result> = [],
 ): boolean => isEmpty(objects) && isEmpty(poeple) && isEmpty(containers);
 
-const sliceResults = (resultsMap: GenericResultMap | null, abTest: ABTest) => {
+const sliceResults = (resultsMap: JiraResultsMap | null, abTest: ABTest) => {
   const { objects, containers, people } = resultsMap
     ? resultsMap
     : DEFAULT_JIRA_RESULTS_MAP;
@@ -55,7 +55,7 @@ const sliceResults = (resultsMap: GenericResultMap | null, abTest: ABTest) => {
 };
 
 export const mapRecentResultsToUIGroups = (
-  recentlyViewedObjects: GenericResultMap | null,
+  recentlyViewedObjects: JiraResultsMap | null,
   searchSessionId: string,
   features: JiraFeatures,
   appPermission?: JiraApplicationPermission,