+++ packages/core/avatar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/avatar
 
+## 16.0.6
+
+- Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):
+  - @atlaskit/docs@8.1.3
+  - @atlaskit/button@13.0.9
+  - @atlaskit/field-base@13.0.6
+  - @atlaskit/section-message@4.0.5
+  - @atlaskit/toggle@7.0.3
+  - @atlaskit/tooltip@15.0.2
+  - @atlaskit/icon@19.0.0
+
 ## 16.0.5
 
 ### Patch Changes