+++ packages/core/modal-dialog/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @atlaskit/modal-dialog
 
+## 10.0.7
+
+- Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):
+  - @atlaskit/docs@8.1.3
+  - @atlaskit/avatar@16.0.6
+  - @atlaskit/button@13.0.9
+  - @atlaskit/checkbox@8.0.5
+  - @atlaskit/field-radio-group@6.0.4
+  - @atlaskit/form@6.1.1
+  - @atlaskit/inline-dialog@12.0.3
+  - @atlaskit/portal@3.0.7
+  - @atlaskit/select@9.1.8
+  - @atlaskit/textfield@2.0.3
+  - @atlaskit/icon@19.0.0
+
 ## 10.0.6
 
 ### Patch Changes