+++ packages/core/nps/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @atlaskit/nps
 
+## 6.0.4
+
+- Updated dependencies [06326ef3f7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/06326ef3f7):
+  - @atlaskit/docs@8.1.3
+  - @atlaskit/button@13.0.9
+  - @atlaskit/checkbox@8.0.5
+  - @atlaskit/dropdown-menu@8.0.8
+  - @atlaskit/icon@19.0.0
+
 ## 6.0.3
 
 ### Patch Changes